Rename useProjectKeyStore to useProjectKey
