+++ stores/index.ts
@@ -1,6 +1,5 @@
-export { default as useProjectKeyStore } from './useProjectKeyStore';
-
 export { default as useApi } from './useApi';
+export { default as useProjectKey } from './useProjectKey';
 export { default as useProject } from './useProject';
 export { default as useResourceDataType } from './useResourceDataType';
 export { default as useResourceData } from './useResourceData';
